+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 This log documents the significant changes for each release.
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [1.15.0] - 2020-10-30
+### Added
+- Build/Load Cohort (instead of Patient selection)
+## Changed
+- Allow patients to be selected by Observation codes which do not have
+  values provided for them
+- Changed the Patient.active parameter input control from a checkbox to
+  a radio button group
+## Fixed
+- Exception that can be thrown after removing a resource type tab
+- Results of not all queries were considered when filtering patient resources
+- Extra commas in http requests for data needed for resource tabs
+
 ## [1.14.0] - 2020-10-16
 ### Added
 - Allowance to choose which resource types users want to see